update code structure
